chore: create log-validator-bulk component #57
