+++ pages/[project]/changes_logs.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
+import { intersection, uniq } from 'underscore'
 import type { Geometry } from 'geojson'
-import type { Log, ObjectId, Project } from '~/libs/types'
+import type { Log, ObjectId, Project, User } from '~/libs/types'
 
 definePageMeta({
   validate({ params }) {
@@ -8,8 +9,8 @@ definePageMeta({
   },
 })
 
-const params = useRoute().params
-const projectSlug = params.project as string
+const route = useRoute()
+const projectSlug = route.params.project as string
 const project = ref<Project>()
 const logs = ref<Log[]>()
 
@@ -31,6 +32,47 @@ catch (err: any) {
   ElMessage.error(err.message)
 }
 
+const logsFiltered = computed(() => {
+  if (!logs.value?.length) {
+    return []
+  }
+
+  return logs.value.filter((log) => {
+    const changesetsUsers
+      = route.query.filterByUsers !== undefined
+      && uniq(
+        (log.base ? log.changesets.slice(1) : log.changesets).map(
+          (changeset) => changeset.user,
+        ),
+      )
+    return (
+      (route.query.filterByAction === undefined
+      || Object.values(log.diff_attribs || {})
+        .concat(Object.values(log.diff_tags || {}))
+        .some(
+          (actions) =>
+            actions?.some(
+              (action) => action[0] === route.query.filterByAction,
+            ) || false,
+        ))
+        && (route.query.filterByUserGroups === undefined
+        || log.matches.some((match) =>
+          match.user_groups.includes(route.query.filterByUserGroups as string),
+        ))
+        && (route.query.filterBySelectors === undefined
+        || log.matches.some((match) =>
+          matchFilterBySelectors(match.selectors),
+        ))
+        && (route.query.filterByUsers === undefined
+        || (changesetsUsers
+        && changesetsUsers.length === 1
+        && changesetsUsers[0] === route.query.filterByUsers))
+        && (route.query.filterByDate === undefined
+        || log.change.created.substring(0, 10) === route.query.filterByDate)
+    )
+  })
+})
+
 const baseGeoms = computed(() => {
   if (!logs.value) {
     return []
@@ -49,6 +91,18 @@ const changeGeoms = computed(() => {
   return logs.value.map((log) => log.change.geom)
 })
 
+const user = useState<User>('user')
+const isProjectUser = computed(() => {
+  if (!user.value) {
+    return false
+  }
+  return !!user.value.projects?.includes(projectSlug)
+})
+
+function matchFilterBySelectors(selectors: string[]): boolean {
+  return !!(route.query.filterBySelectors && intersection(selectors, route.query.filterBySelectors?.toString()).length > 0)
+}
+
 function removeLogs(objectIds: ObjectId[]) {
   logs.value = logs.value?.filter(
     (log) =>
@@ -56,6 +110,42 @@ function removeLogs(objectIds: ObjectId[]) {
         (objectId) => log.objtype === objectId.objtype && log.id === objectId.id,
       ) === -1,
   )
+}
+
+async function validate(objectIds: ObjectId[]) {
+  try {
+    await useFetchWithCache(
+      'accept',
+      `${useRuntimeConfig().public.API}/projects/${projectSlug}/changes_logs/accept`,
+      {
+        credentials: 'include',
+        method: 'POST',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(objectIds),
+      },
+    )
+    removeLogs(objectIds)
+  }
+  catch (err: any) {
+    ElMessage.error(err.message)
+  }
+}
+
+const router = useRouter()
+async function handleBulkValidation() {
+  const objectIds = logsFiltered.value.map((log) => ({
+    objtype: log.objtype,
+    id: log.id,
+    version: log.change.version,
+    deleted: log.change.deleted,
+  }))
+
+  validate(objectIds)
+
+  await router.replace({ ...route, query: undefined })
 }
 </script>
 
@@ -72,6 +162,14 @@ function removeLogs(objectIds: ObjectId[]) {
       />
     </el-row>
     <log-filters />
-    <logs :project-slug="projectSlug" @remove-logs="removeLogs($event)" />
+    <log-validator-bulk
+      v-if="isProjectUser && (Object.keys(route.query).length)"
+      @bulk-validation="handleBulkValidation"
+    />
+    <!-- <logs
+      :logs="logsFiltered"
+      :project-slug="projectSlug"
+      @validate="validate"
+    /> -->
   </el-container>
 </template>
